requirements.txt
# LangChain 相关
langchain>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-openai>=0.2.0,<0.4.0
langgraph>=0.2.0,<0.3.0

# 数据库相关
psycopg2-binary>=2.9.0,<3.0.0
pgvector>=0.3.0,<0.4.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
asyncpg>=0.29.0,<0.30.0

# Web 框架
fastapi>=0.110.0,<0.111.0
uvicorn[standard]>=0.30.0,<0.31.0
sse-starlette>=2.1.0,<3.0.0

# 任务队列
celery[redis]>=5.4.0,<6.0.0
redis>=5.0.0,<6.0.0

# 工具库
pydantic>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
tiktoken>=0.7.0,<0.8.0
numpy>=1.26.0,<2.0.0
